+++ apps/native-component-list/src/navigation/ExpoApisStackNavigator.tsx
@@ -10,11 +10,12 @@ import getStackNavWithConfig from '../navigation/StackConfig';
 import { AudioScreens } from '../screens/Audio/AudioScreen';
 import { BlobScreens } from '../screens/Blob/BlobScreen';
 import { CalendarsScreens } from '../screens/CalendarsScreen';
+import { apiScreensToListElements } from '../screens/ComponentListScreen';
 import { ContactsScreens } from '../screens/Contacts/ContactsScreen';
 import ExpoApis from '../screens/ExpoApisScreen';
 import { MediaLibraryScreens } from '../screens/MediaLibrary@Next/MediaLibraryScreens';
 import { ModulesCoreScreens } from '../screens/ModulesCore/ModulesCoreScreen';
-import { type ScreenApiItem, type ScreenConfig } from '../types/ScreenConfig';
+import { type ScreenConfig } from '../types/ScreenConfig';
 
 const Stack = createNativeStackNavigator();
 
@@ -55,7 +56,7 @@ export const ScreensList: ScreenConfig[] = [
     getComponent() {
       return optionalRequire(() => require('../screens/CellularScreen'));
     },
-    name: 'Cellular',
+    name: 'Cellular (device-only)',
   },
   {
     getComponent() {
@@ -144,7 +145,7 @@ export const ScreensList: ScreenConfig[] = [
     getComponent() {
       return optionalRequire(() => require('../screens/BrightnessScreen'));
     },
-    name: 'Brightness',
+    name: 'Brightness (device-only)',
   },
   {
     getComponent() {
@@ -169,6 +170,12 @@ export const ScreensList: ScreenConfig[] = [
       return optionalRequire(() => require('../screens/FileSystemScreen'));
     },
     name: 'FileSystem',
+  },
+  {
+    getComponent() {
+      return optionalRequire(() => require('../screens/FileSystemLegacyScreen'));
+    },
+    name: 'FileSystem@legacy',
   },
   {
     getComponent() {
@@ -454,11 +461,7 @@ export const Screens: ScreenConfig[] = [
   ...CalendarsScreens,
 ];
 
-export const screenApiItems: ScreenApiItem[] = ScreensList.map(({ name, route }) => ({
-  name,
-  route: '/apis/' + (route ?? name.toLowerCase()),
-  isAvailable: true,
-}));
+export const screenApiItems = apiScreensToListElements(ScreensList);
 
 function ExpoApisStackNavigator(props: { navigation: BottomTabNavigationProp<any> }) {
   const { theme } = useTheme();