+++ packages/@expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
@@ -13,6 +13,7 @@ import type {
   CustomResolutionContext,
 } from '@expo/metro/metro-resolver';
 import { resolve as metroResolver } from '@expo/metro/metro-resolver';
+import type { SourceFileResolution } from '@expo/metro/metro-resolver/types';
 import chalk from 'chalk';
 import fs from 'fs';
 import path from 'path';
@@ -652,19 +653,24 @@ export function withExtendedResolver(
 
       const normalizedPath = normalizeSlashes(result.filePath);
 
-      if (normalizedPath.endsWith('expo-router/build/layouts/_web-modal.js')) {
-        if (env.EXPO_UNSTABLE_WEB_MODAL) {
-          try {
-            const webModal = doResolve('expo-router/build/layouts/ExperimentalModalStack.js');
-            if (webModal.type === 'sourceFile') {
-              debug('Using `_unstable-web-modal` implementation.');
-              return webModal;
-            }
-          } catch (error) {
-            // Fallback to react-navigation web modal implementation.
-          }
+      const doReplace = (from: string, to: string | undefined, options?: { throws?: boolean }) =>
+        doReplaceHelper(from, to, {
+          normalizedPath,
+          doResolve,
+          ...options,
+        });
+      const doReplaceStrict = (from: string, to: string | undefined) =>
+        doReplace(from, to, { throws: true });
+
+      if (env.EXPO_UNSTABLE_WEB_MODAL) {
+        const webModalModule = doReplace(
+          'expo-router/build/layouts/_web-modal.js',
+          'expo-router/build/layouts/ExperimentalModalStack.js'
+        );
+        if (webModalModule) {
+          debug('Using `_unstable-web-modal` implementation.');
+          return webModalModule;
         }
-        debug("Using React Navigation's web modal implementation.");
       }
 
       if (platform === 'web') {
@@ -712,12 +718,31 @@ export function withExtendedResolver(
 
         // Shim out React Native native runtime globals in server mode for native.
         if (isServer) {
-          if (normalizedPath.endsWith('react-native/Libraries/Core/InitializeCore.js')) {
+          const emptyModule = doReplace('react-native/Libraries/Core/InitializeCore.js', undefined);
+          if (emptyModule) {
             debug('Shimming out InitializeCore for React Native in native SSR bundle');
-            return {
-              type: 'empty',
-            };
+            return emptyModule;
           }
+        }
+
+        const hmrModule = doReplaceStrict(
+          'react-native/Libraries/Utilities/HMRClient.js',
+          'expo/src/async-require/hmr.ts'
+        );
+        if (hmrModule) return hmrModule;
+
+        if (env.EXPO_UNSTABLE_LOG_BOX) {
+          const logBoxModule = doReplace(
+            'react-native/Libraries/LogBox/LogBoxInspectorContainer.js',
+            '@expo/log-box/swap-rn-logbox.js'
+          );
+          if (logBoxModule) return logBoxModule;
+
+          const logBoxParserModule = doReplace(
+            'react-native/Libraries/LogBox/Data/parseLogBoxLog.js',
+            '@expo/log-box/swap-rn-logbox-parser.js'
+          );
+          if (logBoxParserModule) return logBoxParserModule;
         }
       }
 
@@ -802,6 +827,47 @@ export function withExtendedResolver(
   return withMetroErrorReportingResolver(
     withMetroSupervisingTransformWorker(metroConfigWithCustomContext)
   );
+}
+
+function doReplaceHelper(
+  from: string,
+  to: string | undefined,
+  {
+    throws = false,
+    normalizedPath,
+    doResolve,
+  }: {
+    throws?: boolean;
+    normalizedPath: string;
+    doResolve: StrictResolver;
+  }
+): SourceFileResolution | { type: 'empty' } | undefined {
+  if (!normalizedPath.endsWith(from)) {
+    return undefined;
+  }
+
+  if (to === undefined) {
+    return {
+      type: 'empty',
+    };
+  }
+
+  try {
+    const hmrModule = doResolve(to);
+    if (hmrModule.type === 'sourceFile') {
+      debug(`Using \`${to}\` implementation.`);
+      return hmrModule;
+    }
+  } catch (resolutionError) {
+    if (throws) {
+      throw new Error(`Failed to replace ${from} with ${to}. Resolution of ${to} failed.`, {
+        cause: resolutionError,
+      });
+    }
+
+    debug(`Failed to resolve ${to} when swapping from ${from}: ${resolutionError}`);
+  }
+  return undefined;
 }
 
 /** @returns `true` if the incoming resolution should be swapped. */