+++ packages/@expo/log-box/metro.config.js
@@ -1,0 +1,9 @@
+// Learn more https://docs.expo.io/guides/customizing-metro
+const { getDefaultConfig } = require('expo/metro-config');
+
+/** @type {import('expo/metro-config').MetroConfig} */
+const config = getDefaultConfig(__dirname);
+
+config.transformerPath = require.resolve('./shadowDomCssTransformer');
+
+module.exports = config;