+++ packages/@expo/metro-config/src/serializer/fork/hmrJSBundle.ts
@@ -10,17 +10,17 @@
  */
 
 import type { DeltaResult, Module, ReadOnlyGraph } from '@expo/metro/metro/DeltaBundler';
-import type { HmrModule } from '@expo/metro/metro-runtime/modules/types.flow';
+import type { HmrModule } from '@expo/metro/metro-runtime/modules/types';
 import { addParamsToDefineCall } from '@expo/metro/metro-transform-plugins';
 import jscSafeUrl from 'jsc-safe-url';
 import path from 'node:path';
-import type { UrlWithParsedQuery as EntryPointURL } from 'node:url';
-import url from 'node:url';
 
 import { isJsModule, wrapModule } from './js';
 
+const debug = require('debug')('Metro:HMR');
+
 type Options = {
-  clientUrl: EntryPointURL;
+  clientUrl: URL;
   createModuleId: (id: string) => number;
   includeAsyncPaths: boolean;
   projectRoot: string;
@@ -36,27 +36,39 @@ function generateModules(
 
   for (const module of sourceModules) {
     if (isJsModule(module)) {
-      // Construct a bundle URL for this specific module only
-      const getURL = (extension: 'bundle' | 'map') => {
-        const moduleUrl = url.parse(url.format(options.clientUrl), true);
-        // the legacy url object is parsed with both "search" and "query" fields.
-        // for the "query" field to be used when formatting the object bach to string, the "search" field must be empty.
-        // https://nodejs.org/api/url.html#urlformaturlobject:~:text=If%20the%20urlObject.search%20property%20is%20undefined
-        moduleUrl.search = '';
-        moduleUrl.pathname = path.relative(
-          options.serverRoot ?? options.projectRoot,
-          path.join(
-            path.dirname(module.path),
-            path.basename(module.path, path.extname(module.path)) + '.' + extension
-          )
+      const getPathname = (extension: 'bundle' | 'map') => {
+        return (
+          path
+            .relative(
+              options.serverRoot ?? options.projectRoot,
+              path.join(
+                path.dirname(module.path),
+                path.basename(module.path, path.extname(module.path)) + '.' + extension
+              )
+            )
+            .split(path.sep)
+            // using this Metro particular convention for encoding file paths as URL paths.
+            .map((segment) => encodeURIComponent(segment))
+            .join('/')
         );
-
-        delete moduleUrl.query.excludeSource;
-        return url.format(moduleUrl);
       };
 
-      const sourceMappingURL = getURL('map');
-      const sourceURL = jscSafeUrl.toJscSafeUrl(getURL('bundle'));
+      const clientUrl = new URL(options.clientUrl);
+      clientUrl.searchParams.delete('excludeSource');
+
+      clientUrl.pathname = getPathname('map');
+      const sourceMappingURL = clientUrl.toString();
+
+      clientUrl.pathname = getPathname('bundle');
+      const sourceURL = jscSafeUrl.toJscSafeUrl(clientUrl.toString());
+
+      debug(
+        'got sourceMappingURL: %s\nand sourceURL: %s\nfor module: %s',
+        sourceMappingURL,
+        sourceURL,
+        module.path
+      );
+
       const code =
         prepareModule(module, graph, options) +
         `\n//# sourceMappingURL=${sourceMappingURL}\n` +
@@ -76,7 +88,7 @@ function generateModules(
 function prepareModule(module: Module<any>, graph: ReadOnlyGraph<any>, options: Options): string {
   const code = wrapModule(module, {
     ...options,
-    sourceUrl: url.format(options.clientUrl),
+    sourceUrl: options.clientUrl.toString(),
     dev: true,
     skipWrapping: false,
     computedAsyncModulePaths: null,