+++ packages/expo/android/src/main/java/expo/modules/ExpoReactHostFactory.kt
@@ -133,7 +133,6 @@ object ExpoReactHostFactory {
           true,
           useDevSupport
         )
-
       hostHandlers.forEach { handler ->
         handler.onDidCreateDevSupportManager(reactHostImpl.devSupportManager)
       }